feat(armor_detector): 添加角点校正参数化配置，修改部分过程变量数据类型，更新todo list。

+++ src/rm_auto_aim/armor_detector/include/armor_detector/light_corner_corrector.hpp
@@ -12,7 +12,7 @@ struct SymmetryAxis
 {
   cv::Point2f centroid;
   cv::Point2f direction;
-  float mean_val;  // Mean brightness
+  double mean_val;  // Mean brightness
 };
 
 // This class is used to improve the precision of the corner points of the light bar.
